Adding microservice commands and merging from fcli/main
