+++ fcli-fod/src/main/java/com/fortify/cli/fod/_main/cli/cmd/FoDCommands.java
@@ -1,8 +1,8 @@
 package com.fortify.cli.fod._main.cli.cmd;
 
 import com.fortify.cli.fod.app.cli.cmd.FoDAppCommands;
-import com.fortify.cli.fod.appmicroservice.cli.cmd.FoDAppMicroserviceCommands;
-import com.fortify.cli.fod.apprelease.cli.cmd.FoDAppRelCommands;
+import com.fortify.cli.fod.microservice.cli.cmd.FoDAppMicroserviceCommands;
+import com.fortify.cli.fod.release.cli.cmd.FoDAppRelCommands;
 import com.fortify.cli.fod.dast_scan.cli.cmd.FoDDastScanCommands;
 import com.fortify.cli.fod.lookup.cli.cmd.FoDLookupCommands;
 import com.fortify.cli.fod.rest.cli.cmd.FoDRestCommands;
