--- fcli-fod/src/main/java/com/fortify/cli/fod/appmicroservice/cli/cmd/FoDAppMicroserviceDeleteCommand.java
@@ -1,85 +1,0 @@
-/*******************************************************************************
- * (c) Copyright 2020 Micro Focus or one of its affiliates
- *
- * Permission is hereby granted, free of charge, to any person obtaining a
- * copy of this software and associated documentation files (the
- * "Software"), to deal in the Software without restriction, including without
- * limitation the rights to use, copy, modify, merge, publish, distribute,
- * sublicense, and/or sell copies of the Software, and to permit persons to
- * whom the Software is furnished to do so, subject to the following
- * conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY
- * KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE
- * WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
- * PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
- * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
- * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
- * IN THE SOFTWARE.
- ******************************************************************************/
-package com.fortify.cli.fod.appmicroservice.cli.cmd;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fortify.cli.common.output.cli.cmd.unirest.IUnirestJsonNodeSupplier;
-import com.fortify.cli.common.output.spi.transform.IActionCommandResultSupplier;
-import com.fortify.cli.common.output.spi.transform.IRecordTransformer;
-import com.fortify.cli.fod.app.cli.mixin.FoDAppResolverMixin;
-import com.fortify.cli.fod.appmicroservice.cli.mixin.FoDAppMicroserviceResolverMixin;
-import com.fortify.cli.fod.appmicroservice.helper.FoDAppMicroserviceDescriptor;
-import com.fortify.cli.fod.appmicroservice.helper.FoDAppMicroserviceHelper;
-import com.fortify.cli.fod.output.cli.AbstractFoDOutputCommand;
-import com.fortify.cli.fod.output.mixin.FoDOutputHelperMixins;
-import io.micronaut.core.annotation.ReflectiveAccess;
-import kong.unirest.UnirestInstance;
-import lombok.Getter;
-import picocli.CommandLine;
-import picocli.CommandLine.Command;
-import picocli.CommandLine.Mixin;
-import picocli.CommandLine.Model.CommandSpec;
-import picocli.CommandLine.Spec;
-
-import java.util.ResourceBundle;
-
-@ReflectiveAccess
-@Command(name = FoDOutputHelperMixins.Delete.CMD_NAME)
-public class FoDAppMicroserviceDeleteCommand extends AbstractFoDOutputCommand implements IUnirestJsonNodeSupplier, IRecordTransformer, IActionCommandResultSupplier {
-    @Getter @Mixin private FoDOutputHelperMixins.Delete outputHelper;
-    @Spec CommandSpec spec;
-    ResourceBundle bundle = ResourceBundle.getBundle("com.fortify.cli.fod.i18n.FoDMessages");
-
-    @Mixin private FoDAppMicroserviceResolverMixin.PositionalParameter appMicroserviceResolver;
-
-    @Override
-    public JsonNode getJsonNode(UnirestInstance unirest) {
-        FoDAppMicroserviceDescriptor appMicroserviceDescriptor;
-        try {
-            appMicroserviceDescriptor = FoDAppMicroserviceHelper.getAppMicroservice(unirest, appMicroserviceResolver.getAppAndMicroserviceNameOrId(), ":", true);
-        } catch (JsonProcessingException e) {
-            throw new CommandLine.ParameterException(spec.commandLine(),
-                    bundle.getString("fcli.fod.microservice.update.invalid-parameter"));
-        }
-
-        return FoDAppMicroserviceHelper.deleteAppMicroservice(unirest, appMicroserviceDescriptor);
-    }
-
-    @Override
-    public JsonNode transformRecord(JsonNode record) {
-        return FoDAppMicroserviceHelper.renameFields(record);
-    }
-
-    @Override
-    public String getActionCommandResult() {
-        return "DELETED";
-    }
-    
-    @Override
-    public boolean isSingular() {
-        return true;
-    }
-}
