+++ fcli-fod/src/main/resources/com/fortify/cli/fod/i18n/FoDMessages.properties
@@ -34,6 +34,7 @@ fcli.fod.app.create.type = The type of application to create. Valid values: ${CO
 fcli.fod.app.create.criticality = The business criticality of the application. Valid values: ${COMPLETION-CANDIDATES}.
 fcli.fod.app.create.status = The SDLC lifecycle status of the release. Valid values: ${COMPLETION-CANDIDATES}
 fcli.fod.app.create.attr = Attribute id or name and its value to set on the application.
+fcli.fod.app.create.skip-if-exists = Check to see if application name already exists before creating.
 
 # For the "fod app delete" command
 fcli.fod.app.delete.usage.description = Delete an application from Fortify on Demand (FoD).
@@ -67,6 +68,7 @@ fcli.fod.release.create.copy-state = Copy the state of an existing release.
 fcli.fod.release.create.copy-release = The id or name of the release to copy.
 fcli.fod.release.create.microservice = The id or name of the microservice to create the release on.
 fcli.fod.release.create.status = SDLC lifecycle status of the release. Valid values: ${COMPLETION-CANDIDATES}.
+fcli.fod.release.create.skip-if-exists = Check to see if application and release name already exists before creating.
 
 # For the "fod release delete" command
 fcli.fod.release.delete.usage.description = Delete an application release from Fortify on Demand (FoD).
@@ -91,6 +93,7 @@ fcli.fod.microservice.usage.description = Commands for interacting with applicat
 # For the "fod microservice create" command
 fcli.fod.microservice.create.usage.description = Create a new application microservice on Fortify on Demand (FoD).
 fcli.fod.microservice.create.name = The name of the microservice to create for the application.
+fcli.fod.microservice.create.skip-if-exists = Check to see if release and microservice name already exists before creating.
 
 # For the "fod microservice delete" command
 fcli.fod.microservice.delete.usage.description = Delete an application microservice from Fortify on Demand (FoD).
